+++ cli/prev.c
@@ -34,10 +34,10 @@ int tman_cli_prev(int argc, char **argv, struct tman_context *ctx)
         if (quiet == FALSE)
             elog(status, errfmt, tman_strerror());
         return status;
-    } else if (tmancfg->usehooks == TRUE &&
-               (status =
-                tman_hook_action(ctx, tmancfg->hooks, &args,
-                                 "prev")) != LIBTMAN_OK) {
+    } else if (tmancfg->usehooks == TRUE && (status =
+                                             tman_hook_action(&args,
+                                                              "prev")) !=
+               LIBTMAN_OK) {
         if (quiet == FALSE)
             elog(status, errfmt, tman_strerror());
         return status;